refactor: reorganize file structure and update import paths
