+++ src/helper/checkCancel.ts
