+++ vitest.config.ts
@@ -8,7 +8,7 @@ export default defineConfig(({ mode }) => {
   return {
     plugins: [viteTsconfigPaths()],
     test: {
-      // setupFiles: 'vitest.setup.ts',
+      setupFiles: 'vitest.setup.ts',
       clearMocks: true,
       reporters: ['default', 'junit', 'vitest-sonar-reporter'],
       outputFile: {
